a bit refactor and change project structure to avoid merge conflict
