+++ src/main/java/com/robosh/ejournal/data/dto/director/DirectorDTO.java
@@ -1,0 +1,26 @@
+package com.robosh.ejournal.data.dto.director;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class DirectorDTO {
+
+    protected Long id;
+
+    protected String firstName;
+
+    protected String secondName;
+
+    protected String lastName;
+
+    protected String email;
+
+    protected String phone;
+}
