+++ src/main/java/com/robosh/ejournal/data/entity/Director.java
@@ -1,0 +1,8 @@
+package com.robosh.ejournal.data.entity;
+
+import javax.persistence.Entity;
+
+@Entity(name = "director")
+public class Director extends User {
+
+}
