+++ src/main/java/com/robosh/ejournal/data/entity/Subject.java
@@ -1,0 +1,24 @@
+package com.robosh.ejournal.data.entity;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.validation.constraints.NotBlank;
+
+@Data
+@NoArgsConstructor
+@Entity(name = "subject")
+public class Subject {
+
+    @Id
+    @GeneratedValue
+    private Long id;
+
+    @NotBlank
+    @Column(name = "name")
+    private String name;
+}
