+++ src/main/java/com/robosh/ejournal/data/entity/admin/AdminRole.java
@@ -1,0 +1,5 @@
+package com.robosh.ejournal.data.entity.admin;
+
+public enum AdminRole {
+    ADMIN, SUPER_ADMIN
+}
