+++ src/main/java/com/robosh/ejournal/data/mapping/admin/AdminDTOMapper.java
@@ -1,0 +1,7 @@
+package com.robosh.ejournal.data.mapping.admin;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class AdminDTOMapper {
+}
