+++ src/main/java/com/robosh/ejournal/data/mapping/admin/AdminMapper.java
@@ -1,0 +1,11 @@
+package com.robosh.ejournal.data.mapping.admin;
+
+import com.robosh.ejournal.data.dto.admin.AdminInfoDTO;
+import com.robosh.ejournal.data.entity.admin.Admin;
+import org.mapstruct.Mapper;
+
+@Mapper(uses = {AdminDTOMapper.class}, componentModel = "spring")
+public interface AdminMapper {
+
+    AdminInfoDTO fromAdminToAdminInfoTDO(Admin admin);
+}
