+++ src/main/java/com/robosh/ejournal/data/repository/ParentRepository.java
@@ -1,0 +1,8 @@
+package com.robosh.ejournal.data.repository;
+
+import com.robosh.ejournal.data.entity.Parent;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ParentRepository extends JpaRepository<Parent, Long> {
+
+}
