+++ src/main/java/com/robosh/ejournal/data/repository/StudentRepository.java
@@ -1,0 +1,8 @@
+package com.robosh.ejournal.data.repository;
+
+import com.robosh.ejournal.data.entity.Student;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface StudentRepository extends JpaRepository<Student, Long> {
+
+}
