+++ src/main/java/com/robosh/ejournal/data/repository/TeacherRepository.java
@@ -1,0 +1,8 @@
+package com.robosh.ejournal.data.repository;
+
+import com.robosh.ejournal.data.entity.Teacher;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TeacherRepository extends JpaRepository<Teacher, Long> {
+
+}
