--- src/main/java/com/robosh/ejournal/entity/Director.java
@@ -1,8 +1,0 @@
-package com.robosh.ejournal.entity;
-
-import javax.persistence.Entity;
-
-@Entity(name = "director")
-public class Director extends User {
-
-}
