--- src/main/java/com/robosh/ejournal/entity/Schedule.java
@@ -1,39 +1,0 @@
-package com.robosh.ejournal.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.JoinTable;
-import javax.persistence.OneToMany;
-import javax.persistence.OneToOne;
-import javax.validation.constraints.NotNull;
-import java.util.List;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Entity(name = "schedule")
-public class Schedule {
-
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    @NotNull
-    @OneToOne
-    @JoinColumn(name = "group_id")
-    private Group group;
-
-    @OneToMany
-    @JoinTable(name = "schedule_schedule_item",
-            joinColumns = @JoinColumn(name = "schedule_id"),
-            inverseJoinColumns = @JoinColumn(name = "schedule_item_id"))
-    private List<ScheduleItem> scheduleList;
-}
