--- src/main/java/com/robosh/ejournal/entity/Subject.java
@@ -1,24 +1,0 @@
-package com.robosh.ejournal.entity;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.validation.constraints.NotBlank;
-
-@Data
-@NoArgsConstructor
-@Entity(name = "subject")
-public class Subject {
-
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    @NotBlank
-    @Column(name = "name")
-    private String name;
-}
