--- src/main/java/com/robosh/ejournal/entity/admin/Admin.java
@@ -1,57 +1,0 @@
-package com.robosh.ejournal.entity.admin;
-
-import com.robosh.ejournal.entity.School;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.lang.Nullable;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Pattern;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Entity(name = "Admin")
-public class Admin {
-
-    @Id
-    @GeneratedValue
-    @Column(name = "id")
-    private Long id;
-
-    @NotBlank
-    @Column(name = "first_name", nullable = false)
-    private String firstName;
-
-    @NotBlank
-    @Column(name = "last_name", nullable = false)
-    private String lastName;
-
-    @Pattern(regexp = "[A-Za-z0-9+_.-]+@[a-z.-]+\\.[a-z]{2,8}")
-    @Column(name = "email", nullable = false, unique = true)
-    private String email;
-
-    @NotBlank
-    @Column(name = "password", nullable = false)
-    private String password;
-
-    @Nullable
-    @ManyToOne
-    @JoinColumn(name = "school_id", nullable = false)
-    private School school;
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "role", nullable = false)
-    private AdminRole adminRole;
-}
