--- src/main/java/com/robosh/ejournal/entity/admin/AdminRole.java
@@ -1,5 +1,0 @@
-package com.robosh.ejournal.entity.admin;
-
-public enum AdminRole {
-    ADMIN, SUPER_ADMIN
-}
