--- src/main/java/com/robosh/ejournal/mapper/DirectorMapper.java
@@ -1,14 +1,0 @@
-package com.robosh.ejournal.mapper;
-
-import com.robosh.ejournal.dto.DirectorDTO;
-import com.robosh.ejournal.entity.Director;
-import org.mapstruct.Mapper;
-import org.mapstruct.factory.Mappers;
-
-@Mapper(componentModel = "spring")
-public interface DirectorMapper {
-
-    DirectorMapper directorMapper = Mappers.getMapper(DirectorMapper.class);
-
-    DirectorDTO fromDirectorToDirectorDTO(Director director);
-}
