--- src/main/java/com/robosh/ejournal/repository/AdminRepository.java
@@ -1,10 +1,0 @@
-package com.robosh.ejournal.repository;
-
-import com.robosh.ejournal.entity.admin.Admin;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface AdminRepository extends JpaRepository<Admin, Long> {
-    List<Admin> findAll();
-}
