--- src/main/java/com/robosh/ejournal/repository/ParentRepository.java
@@ -1,8 +1,0 @@
-package com.robosh.ejournal.repository;
-
-import com.robosh.ejournal.entity.Parent;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ParentRepository extends JpaRepository<Parent, Long> {
-
-}
