--- src/main/java/com/robosh/ejournal/repository/StudentRepository.java
@@ -1,8 +1,0 @@
-package com.robosh.ejournal.repository;
-
-import com.robosh.ejournal.entity.Student;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StudentRepository extends JpaRepository<Student, Long> {
-
-}
