--- src/main/java/com/robosh/ejournal/repository/TeacherRepository.java
@@ -1,8 +1,0 @@
-package com.robosh.ejournal.repository;
-
-import com.robosh.ejournal.entity.Teacher;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TeacherRepository extends JpaRepository<Teacher, Long> {
-
-}
